修复 PersonMapper 编译错误：移除注解中不兼容的 XML 脚本批量更新

+++ src/main/java/com/prd/controller/RelationTypeDictController.java
@@ -22,7 +22,7 @@ public class RelationTypeDictController {
     }
 
     @PostMapping
-    public Result<Long> save(@RequestBody RelationTypeDict dict) {
+    public Result<Integer> save(@RequestBody RelationTypeDict dict) {
         relationTypeDictService.save(dict);
         return Result.success(dict.getId());
     }
